tests/wpt/tests/css/css-rhythm/parsing/block-step-round-valid.html

<!DOCTYPE html>
<html>
<head>
<title>CSS Rhythm: block-step-round valid values</title>
<link rel="author" title="Tim Nguyen" href="https://github.com/nt1m">
<link rel="help" href="https://drafts.csswg.org/css-rhythm/#block-step-round">
<meta name="assert" content="Parsing valid values for block-step-round property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
    test_valid_value("block-step-round", "up");
    test_valid_value("block-step-round", "down");
    test_valid_value("block-step-round", "nearest");
</script>
</body>
</html>
